ugh read extended
houghlines doesn't really work for finding the homography. I don't know
what the points seemed to be pretty stable and correct. I think we
should just overlap the virtual tictactoe board with the physical one
during calibration.

ughhhh added quadrant extractor. don't wanna test it though

+++ kmeans.py
@@ -78,7 +78,7 @@ def tictactoeMeans(data,k,maxIter):
             centroid = np.mean(data[categories == n, :],0);
             centroids[n,:] = centroid
         if np.array_equal(centroids,oldCentroids): break
-    return centroids
+    return np.float32(centroids)
 
 a = np.zeros((4,2))
 a[0,:] = [1,2.5]
